+++ src/index.ts
@@ -23,10 +23,10 @@ import {
   Camera,
   PerspectiveCamera,
   WebGLRenderer,
+  Texture
 } from 'three';
 
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
-import { GLTFLoader as RDGLTFLoader } from './RDGLTFLoader';
 import { DRACOLoader } from 'three/examples/jsm/loaders/DRACOLoader.js';
 import { KTX2Loader } from 'three/examples/jsm/loaders/KTX2Loader.js';
 
@@ -43,7 +43,6 @@ const grayscale = Gradients.GRAYSCALE;
 const grayscaleTexture = typeof document != 'undefined' ? Util.generateGradientTexture(grayscale) : null;
 
 const defaultOptions: LoaderOptions = {
-  initialTransform: new Matrix4(),
   throttleRequests: true,
   maxRequests: 64,
   updateInterval: 0.1,
@@ -62,8 +61,7 @@ const defaultOptions: LoaderOptions = {
   dracoDecoderPath: null,
   material: null,
   computeNormals: false,
-  shaderCallback: null,
-  loadersGlGltf: true
+  shaderCallback: null
 };
 
 /** 3D Tiles Loader */
@@ -109,27 +107,6 @@ class Loader3DTiles {
       //root.add(tileBoxes)
     }
 
-    // transformations
-    let threeMat = new Matrix4();
-
-    // TODO: This handles the situation where not the root tile is transformed, but one child below the root. There should be a more generic way to handle this...
-    const tileTrasnform = tilesetJson.root.transform
-      ? new Matrix4().fromArray(tilesetJson.root.transform)
-      : new Matrix4();
-    if (tilesetJson.root.children.length == 1 && tilesetJson.root.children[0].transform) {
-      const childTransform = new Matrix4().fromArray(tilesetJson.root.children[0].transform);
-      tileTrasnform.multiply(childTransform);
-    }
-
-    // TODO: Originally the tileset is moved by loaders.gl to its WGS84 matching coordiate. In here, we negate that and bring it back to 0,0,0 with an optional initial transform. If we want to combine the tileset with other geographic layers we might need to go back to those original coordiates
-    threeMat.copy(tileTrasnform).invert();
-
-    const resetTransform = threeMat.clone();
-
-    threeMat.premultiply(options.initialTransform);
-
-    let modelMatrix = new MathGLMatrix4(threeMat.toArray());
-
     const pointcloudUniforms = {
       pointSize: { type: 'f', value: 1.0 },
       gradient: { type: 't', value: gradientTexture },
@@ -146,17 +123,18 @@ class Loader3DTiles {
     let cameraReference = null;
     let rendererReference = null;
 
-    const gltfLoader = options.loadersGlGltf ? new RDGLTFLoader() : new GLTFLoader();
+    const gltfLoader = new GLTFLoader();
 
     if (options.basisTranscoderPath) {
       const ktx2Loader = new KTX2Loader();
       ktx2Loader.detectSupport(props.renderer);
       ktx2Loader.setTranscoderPath(options.basisTranscoderPath + '/');
+      ktx2Loader.setWorkerLimit(1);
 
       gltfLoader.setKTX2Loader(ktx2Loader);
     }
 
-    if (!options.loadersGlGltf && options.dracoDecoderPath) {
+    if (options.dracoDecoderPath) {
       const dracoLoader = new DRACOLoader();
       dracoLoader.setDecoderPath(options.dracoDecoderPath + '/');
       dracoLoader.setWorkerLimit(options.maxConcurrency);
@@ -170,7 +148,6 @@ class Loader3DTiles {
     });
 
     const tileOptions = {
-      modelMatrix: modelMatrix,
       maximumMemoryUsage: options.maximumMemoryUsage,
       maximumScreenSpaceError: options.maximumScreenSpaceError,
       viewDistanceScale: options.viewDistanceScale,
@@ -187,7 +164,7 @@ class Loader3DTiles {
           }
           case TILE_TYPE.SCENEGRAPH:
           case TILE_TYPE.MESH: {
-            tileContent = await createGLTFNodes(gltfLoader, tile, unlitMaterial, options);
+            tileContent = await createGLTFNodes(gltfLoader, tile, unlitMaterial, options, rootTransformInverse);
             break;
           }
           default:
@@ -227,19 +204,46 @@ class Loader3DTiles {
           loadImages: false,
         },
         '3d-tiles': {
-          loadGLTF: options.loadersGlGltf,
+          loadGLTF: false
         },
       },
     });
+    //
+    // transformations
+    let threeMat = new Matrix4();
+
+    const tileTrasnform = new Matrix4();
+
+    if (tileset.root.transform) {
+      tileTrasnform.multiply(new Matrix4().fromArray(tileset.root.transform));
+    }
+    if (tileset.root.children.length == 1 && tileset.root.children[0].transform) {
+        const childTransform = new Matrix4().fromArray(tileset.root.children[0].transform);
+        tileTrasnform.multiply(childTransform);
+    }
+    if (tileTrasnform.equals(new Matrix4().identity()) && tileset.root.header.boundingVolume) {
+      if (tileset.root.header.boundingVolume.region) {
+        // TODO: Handle region type bounding volumes
+        console.warn("Cannot apply a model matrix to bounding volumes of type region. Tileset stays in original geo-coordinates.")
+      } else {
+        tileTrasnform.setPosition(new Vector3(...tileset.root.boundingVolume.center));
+      }
+    }
+
+    // TODO: Originally the tileset is moved by loaders.gl to its WGS84 matching coordiate. In here, we negate that and bring it back to 0,0,0 with an optional initial transform. If we want to combine the tileset with other geographic layers we might need to go back to those original coordiates
+    threeMat.copy(tileTrasnform).invert();
+
+    const resetTransform = threeMat.clone();
+
+    let modelMatrix = new MathGLMatrix4(threeMat.toArray());
+    tileset.modelMatrix = modelMatrix;
 
     let disposeFlag = false;
 
-    const rootCenter = new Vector3().setFromMatrixPosition(options.initialTransform);
+    const rootCenter = new Vector3();
 
     pointcloudUniforms.rootCenter.value.copy(rootCenter);
-    pointcloudUniforms.rootNormal.value.copy(new Vector3(0, 0, 1).applyMatrix4(options.initialTransform).normalize());
-
-    root.applyMatrix4(options.initialTransform);
+    pointcloudUniforms.rootNormal.value.copy(new Vector3(0, 0, 1).normalize());
 
     // Extra stats
     tileset.stats.get('Loader concurrency').count = options.maxConcurrency
@@ -256,6 +260,7 @@ class Loader3DTiles {
     let sseDenominator = null;
 
     const lastRootTransform:Matrix4 = new Matrix4().copy(root.matrixWorld)
+    const rootTransformInverse = new Matrix4().copy(lastRootTransform).invert();
 
     function tilesetUpdate(tileset, renderMap, renderer, camera) {
       if (disposeFlag) {
@@ -427,6 +432,12 @@ class Loader3DTiles {
               lastRootTransform.copy(root.matrixWorld);
               threeMat = resetTransform.clone();
               threeMat.premultiply(lastRootTransform);
+              
+              const rootCenter = new Vector3().setFromMatrixPosition(lastRootTransform);
+              pointcloudUniforms.rootCenter.value.copy(rootCenter);
+              pointcloudUniforms.rootNormal.value.copy(new Vector3(0, 0, 1).applyMatrix4(lastRootTransform).normalize());
+
+              rootTransformInverse.copy(lastRootTransform).invert(); 
               modelMatrix = new MathGLMatrix4(threeMat.toArray());
               tileset.modelMatrix = modelMatrix;
             }
@@ -465,22 +476,36 @@ class Loader3DTiles {
   }
 }
 
-async function createGLTFNodes(gltfLoader, tile, unlitMaterial, options): Promise<Object3D> {
+async function createGLTFNodes(gltfLoader, tile, unlitMaterial, options, rootTransformInverse): Promise<Object3D> {
   return new Promise((resolve, reject) => {
     const rotateX = new Matrix4().makeRotationAxis(new Vector3(1, 0, 0), Math.PI / 2);
+    const shouldRotate = tile.tileset.asset?.gltfUpAxis !== "Z";
+
+    // The computed trasnform already contains the root's transform, so we have to invert it
+    const contentTransform = new Matrix4().fromArray(tile.computedTransform).premultiply(rootTransformInverse);
+
+    if (shouldRotate) {
+      contentTransform.multiply(rotateX); // convert from GLTF Y-up to Z-up
+    }
+
     gltfLoader.parse(
-      options.loadersGlGltf ? tile.content.gltf : tile.content.gltfArrayBuffer,
-      '',
+      tile.content.gltfArrayBuffer,
+      tile.contentUrl ? tile.contentUrl.substr(0,tile.contentUrl.lastIndexOf('/') + 1) : '',
       (gltf) => {
-        const tileContent = gltf.scenes[0].children[0] as Object3D;
-        tileContent.applyMatrix4(rotateX); // convert from GLTF Y-up to Z-up
+        const tileContent = gltf.scenes[0] as Group;
+        tileContent.applyMatrix4(contentTransform); 
+
         tileContent.traverse((object) => {
           if (object instanceof Mesh) {
-            const originalMap = (object.material as MeshStandardMaterial).map;
+            const originalMaterial = (object.material as MeshStandardMaterial);
+            const originalMap = originalMaterial.map;
+
             if (options.material) {
               object.material = options.material.clone();
+              originalMaterial.dispose();
             } else if (options.shading == Shading.FlatTexture) {
               object.material = unlitMaterial.clone();
+              originalMaterial.dispose();
             }
 
             if (options.shading != Shading.ShadedNoTexture) {
@@ -490,6 +515,9 @@ async function createGLTFNodes(gltfLoader, tile, unlitMaterial, options): Promis
                 object.material.map = originalMap;
               }
             } else {
+              if (originalMap) {
+                originalMap.dispose();
+              }
               object.material.map = null;
             }
 
@@ -553,9 +581,23 @@ function createPointNodes(tile, pointcloudUniforms) {
   const tileContent = new Points(geometry, pointcloudMaterial);
   if (d.rtc_center) {
     const c = d.rtc_center;
-    tileContent.applyMatrix4(new Matrix4().makeTranslation(c[0], c[1], c[2]));
+
+    // TODO: In the case of entwine/region bounding volume the modelMatrix also needs to be applied?
+    tileContent.applyMatrix4(new Matrix4().makeTranslation(c[0], c[1], c[2]));//.multiply(threeMat));
   }
   return tileContent;
+}
+
+
+function disposeMaterial(material) {
+
+  if ((material as ShaderMaterial)?.uniforms?.map) {
+    ((material as ShaderMaterial)?.uniforms?.map.value as Texture)?.dispose();
+  }
+  else if (material.map) {
+    (material.map as Texture)?.dispose();
+  }
+  material.dispose();
 }
 
 function disposeNode(node) {
@@ -564,10 +606,12 @@ function disposeNode(node) {
       object.geometry.dispose();
 
       if (object.material.isMaterial) {
-        object.material.dispose();
+        disposeMaterial(object.material);   
       } else {
         // an array of materials
-        for (const material of object.material) material.dispose();
+        for (const material of object.material) {
+          disposeMaterial(material);
+        } 
       }
     }
   });